Move kubecfg utils from minikube to util.
We'll need to also write a kubecfg inside the VM going forward as we
remove the InsecureServing path.
