+++ pkg/minikube/constants/constants.go
@@ -29,7 +29,6 @@ import (
 
 // APIServerPort is the port that the API server should listen on.
 const (
-	APIServerPort    = 8443
 	APIServerName    = "minikubeCA"
 	ClusterDNSDomain = "cluster.local"
 )
